- rewritten service/executor backend - added support for update/creation of nodes - fixed some problems in mylocationmanager - reworked wheelchair state texts in german an added english translation from the website - add database tag to identify update/creation of nodes - reworked distance formatter

+++ Wheelmap/src/wheelmap/org/request/NodeRequestBuilder.java
@@ -7,9 +7,9 @@ package wheelmap.org.request;
 public class NodeRequestBuilder extends RequestBuilder {
 
 	private static final String RESOURCE = "nodes";
-	private int id;
+	private long id;
 
-	public NodeRequestBuilder(final String server, final String apiKey, final AcceptType acceptType, int id) {
+	public NodeRequestBuilder(final String server, final String apiKey, final AcceptType acceptType, long id) {
 		super(server, apiKey, acceptType);
 		this.id=id;
 	}
